Revert "Add actuator specific ObjectMapper"
See gh-12951
See gh-20291
